+++ src/components/RoomDetail/Date/DateContainer.jsx
@@ -8,25 +8,13 @@ import CustomCalendar from "./CustomCalendar";
 import { useRecoilValue, useRecoilState } from "recoil";
 import { selectedDateState } from "../../../recoil/atoms/selectedDate";
 import { availableTimeState } from "../../../recoil/atoms/availableTimeState";
+import useFetchAvailTimes from "./useFetchAvailTimes";
 
 const DateContainer = () => {
   const [selectedDate, setSelectedDate] = useRecoilState(selectedDateState);
-  const [availableTime, setAvailableTime] = useRecoilState(availableTimeState);
-  const formattedDate2 = selectedDate.toISOString().split("T")[0]; //yyyy-mm--dd로 포맷
-  const roomId = useParams();
-  const token = ``;
-
-  const date = useRecoilValue(selectedDateState);
   const [show, setShow] = useState(false);
-  const handleShowCalendar = () => {
-    setShow(!show);
-  };
-
-  const year = date.getFullYear();
-  const month = date.getMonth() + 1;
-  const day = date.getDate();
-
-  const formattedDate = `${year}년 ${month}월 ${day}일 (${getDayOfWeek(date)})`;
+  const roomId = useParams();
+  const date = useRecoilValue(selectedDateState);
 
   const handleDateChange = useCallback(
     (newDate) => {
@@ -34,35 +22,22 @@ const DateContainer = () => {
     },
     [setSelectedDate]
   );
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        // 선택한 날짜를 서버로 전송하는 비동기 함수 호출
-
-        const res = await axios.get(
-          `/partyroom/reservation/search/available?partyroomId=${Number(roomId.roomId)}&date=${formattedDate2}`,
-          {
-            headers: {
-              Authorization: `Bearer ${token}`,
-            },
-          }
-        );
-        // 서버로부터 받은 데이터를 상태에 업데이트
-        setAvailableTime(res.data.availableTimes);
-        console.log(res.data.availableTimes);
-      } catch (error) {
-        console.error("서버 요청 중 오류 발생:", error);
-      }
-    };
-
-    fetchData(); // selectedDate가 변경될 때마다 fetchData 함수 호출
-  }, [selectedDate]);
+  const handleShowCalendar = () => {
+    setShow(!show);
+  };
 
   function getDayOfWeek(date) {
     const daysOfWeek = ["일", "월", "화", "수", "목", "금", "토"];
     return daysOfWeek[date.getDay()];
   }
+
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  const day = date.getDate();
+
+  const formattedDate = `${year}년 ${month}월 ${day}일 (${getDayOfWeek(date)})`;
+
+  useFetchAvailTimes(roomId, selectedDate, date);
 
   return (
     <>
